+++ src/deps/rapidjson/AssetsLoader.cpp
@@ -15,32 +15,32 @@ Assets AssetsLoader::initWithJSONFile(std::string filePath)
     FileData jsonData = ASSET_HANDLER.loadAsset(filePath);
     Assets ret = initWithJSON((const char*)jsonData._data);
     ASSET_HANDLER.unloadAsset(jsonData);
-    
+
     return ret;
 }
 
 Assets AssetsLoader::initWithJSON(const char* json)
 {
     Assets ret;
-    
+
     using namespace rapidjson;
-    
+
     Document d;
     d.Parse<kParseStopWhenDoneFlag>(json);
     assert(d.IsObject());
-    
+
     std::vector<uint16_t> soundIDsAdded;
     if (d.HasMember("music"))
     {
         Value& v = d["music"];
         assert(v.IsObject());
-        
+
         std::string filePath = RapidJSONUtil::getString(v, "filePath");
         uint16_t soundID = 1337;
         ret._soundDescriptors.emplace_back(soundID, filePath, 1);
         soundIDsAdded.emplace_back(soundID);
     }
-    
+
     if (d.HasMember("sounds"))
     {
         Value& v = d["sounds"];
@@ -49,18 +49,18 @@ Assets AssetsLoader::initWithJSON(const char* json)
         {
             const Value& iv = v[i];
             assert(iv.IsObject());
-            
+
             uint16_t soundID = RapidJSONUtil::getUInt(iv, "soundID");
             std::string filePath = RapidJSONUtil::getString(iv, "filePath");
             uint8_t numInstances = RapidJSONUtil::getUInt(iv, "numInstances");
-            
+
             assert(std::find(soundIDsAdded.begin(), soundIDsAdded.end(), soundID) == soundIDsAdded.end());
-            
+
             ret._soundDescriptors.emplace_back(soundID, filePath, numInstances);
             soundIDsAdded.emplace_back(soundID);
         }
     }
-    
+
     if (d.HasMember("shaders"))
     {
         Value& v = d["shaders"];
@@ -69,18 +69,18 @@ Assets AssetsLoader::initWithJSON(const char* json)
         {
             const Value& iv = v[i];
             assert(iv.IsObject());
-            
+
             std::string name = RapidJSONUtil::getString(iv, "name");
             std::string vertexShaderFilePath = RapidJSONUtil::getString(iv, "vertexShaderFilePath");
             std::string fragmentShaderFilePath = RapidJSONUtil::getString(iv, "fragmentShaderFilePath");
-            
+
             ret._shaderDescriptors.emplace_back(name, vertexShaderFilePath, fragmentShaderFilePath);
-            
+
             if (iv.HasMember("uniforms"))
             {
                 ShaderDescriptor& sd = ret._shaderDescriptors.back();
                 std::vector<ShaderUniform>& uniforms = sd._uniforms;
-                
+
                 const Value& v = iv["uniforms"];
                 assert(v.IsArray());
                 for (SizeType i = 0; i < v.Size(); ++i)
@@ -88,16 +88,16 @@ Assets AssetsLoader::initWithJSON(const char* json)
                     const Value& iv = v[i];
                     std::string name = RapidJSONUtil::getString(iv, "name");
                     std::string type = RapidJSONUtil::getString(iv, "type");
-                    
+
                     uniforms.emplace_back(name, type);
                 }
             }
-            
+
             if (iv.HasMember("attributes"))
             {
                 ShaderDescriptor& sd = ret._shaderDescriptors.back();
                 std::vector<ShaderAttribute>& attributes = sd._attributes;
-                
+
                 const Value& v = iv["attributes"];
                 assert(v.IsArray());
                 for (SizeType i = 0; i < v.Size(); ++i)
@@ -106,13 +106,13 @@ Assets AssetsLoader::initWithJSON(const char* json)
                     std::string name = RapidJSONUtil::getString(iv, "name");
                     std::string type = RapidJSONUtil::getString(iv, "type");
                     int count = RapidJSONUtil::getInt(iv, "count", 1);
-                    
+
                     attributes.emplace_back(name, type, count);
                 }
             }
         }
     }
-    
+
     if (d.HasMember("textures"))
     {
         Value& v = d["textures"];
@@ -121,7 +121,7 @@ Assets AssetsLoader::initWithJSON(const char* json)
         {
             const Value& iv = v[i];
             assert(iv.IsObject());
-            
+
             std::string name = RapidJSONUtil::getString(iv, "name");
             std::string normalMapName = RapidJSONUtil::getString(iv, "normalMapName");
             std::string filePath = RapidJSONUtil::getString(iv, "filePath");
@@ -130,44 +130,44 @@ Assets AssetsLoader::initWithJSON(const char* json)
             std::string filterMag = RapidJSONUtil::getString(iv, "filterMag", "SHARP");
             assert(filterMag == "SHARP" || filterMag == "SMOOTH");
             bool mipMap = RapidJSONUtil::getBool(iv, "mipMap", true);
-            
+
             ret._textureDescriptors.emplace_back(name, normalMapName, filePath, filterMin, filterMag, mipMap);
-            
+
             if (iv.HasMember("mappings"))
             {
                 int textureWidth = RapidJSONUtil::getInt(iv, "textureWidth", 2048);
                 int textureHeight = RapidJSONUtil::getInt(iv, "textureHeight", 2048);
-                
+
                 TextureDescriptor& td = ret._textureDescriptors.back();
                 std::map<std::string, Animation>& animations = td._animations;
                 std::map<std::string, TextureRegion>& textureRegions = td._textureRegions;
-                
+
                 const Value& v = iv["mappings"];
                 assert(v.IsObject());
                 for (Value::ConstMemberIterator i = v.MemberBegin(); i != v.MemberEnd(); ++i)
                 {
                     const Value& iv = i->value;
                     assert(iv.IsObject());
-                    
+
                     std::string key = i->name.GetString();
                     int x = RapidJSONUtil::getInt(iv, "x");
                     int y = RapidJSONUtil::getInt(iv, "y");
                     int regionWidth = RapidJSONUtil::getInt(iv, "regionWidth");
                     int regionHeight = RapidJSONUtil::getInt(iv, "regionHeight");
-                    
+
                     if (iv.HasMember("frameTimes") || iv.HasMember("frameTime"))
                     {
                         auto q = animations.find(key);
                         assert(q == animations.end());
-                        
+
                         bool looping = RapidJSONUtil::getBool(iv, "looping", true);
                         int firstLoopingFrame = RapidJSONUtil::getInt(iv, "firstLoopingFrame");
                         int xPadding = RapidJSONUtil::getInt(iv, "xPadding");
                         int yPadding = RapidJSONUtil::getInt(iv, "yPadding");
-                        
+
                         std::vector<uint16_t> frameTimes;
                         size_t numFrames;
-                        
+
                         if (iv.HasMember("frameTimes"))
                         {
                             const Value& va = iv["frameTimes"];
@@ -177,24 +177,24 @@ Assets AssetsLoader::initWithJSON(const char* json)
                                 const Value& iva = va[i];
                                 frameTimes.push_back(iva.GetUint());
                             }
-                            
+
                             numFrames = frameTimes.size();
                         }
                         else
                         {
                             uint16_t frameTime = RapidJSONUtil::getUInt(iv, "frameTime");
                             numFrames = RapidJSONUtil::getUInt(iv, "numFrames");
-                            
+
                             frameTimes.reserve(numFrames);
-                            for (int i = 0; i < numFrames; ++i)
+                            for (size_t i = 0; i < numFrames; ++i)
                             {
                                 frameTimes.push_back(frameTime);
                             }
                         }
-                        
+
                         std::vector<uint16_t> regionWidths;
                         std::vector<uint16_t> regionHeights;
-                        
+
                         if (iv.HasMember("regionWidths"))
                         {
                             const Value& va = iv["regionWidths"];
@@ -208,12 +208,12 @@ Assets AssetsLoader::initWithJSON(const char* json)
                         else
                         {
                             regionWidths.reserve(numFrames);
-                            for (int i = 0; i < numFrames; ++i)
+                            for (size_t i = 0; i < numFrames; ++i)
                             {
                                 regionWidths.push_back(regionWidth);
                             }
                         }
-                        
+
                         if (iv.HasMember("regionHeights"))
                         {
                             const Value& va = iv["regionHeights"];
@@ -227,12 +227,12 @@ Assets AssetsLoader::initWithJSON(const char* json)
                         else
                         {
                             regionHeights.reserve(numFrames);
-                            for (int i = 0; i < numFrames; ++i)
+                            for (size_t i = 0; i < numFrames; ++i)
                             {
                                 regionHeights.push_back(regionHeight);
                             }
                         }
-                        
+
                         uint16_t animationWidth = STLUtil::sum(regionWidths);
                         uint16_t animationHeight = regionHeight;
                         if (iv.HasMember("animationWidth") && iv.HasMember("animationHeight"))
@@ -240,7 +240,7 @@ Assets AssetsLoader::initWithJSON(const char* json)
                             animationWidth = RapidJSONUtil::getUInt(iv, "animationWidth");
                             animationHeight = RapidJSONUtil::getUInt(iv, "animationHeight");
                         }
-                        
+
                         animations.emplace(std::piecewise_construct,
                                            std::forward_as_tuple(key),
                                            std::forward_as_tuple(x, y, regionWidths, regionHeights, animationWidth, animationHeight, textureWidth, textureHeight, looping, firstLoopingFrame, xPadding, yPadding, frameTimes)
@@ -250,7 +250,7 @@ Assets AssetsLoader::initWithJSON(const char* json)
                     {
                         auto q = textureRegions.find(key);
                         assert(q == textureRegions.end());
-                        
+
                         textureRegions.emplace(std::piecewise_construct,
                                                std::forward_as_tuple(key),
                                                std::forward_as_tuple(x, y, regionWidth, regionHeight, textureWidth, textureHeight)
@@ -260,6 +260,6 @@ Assets AssetsLoader::initWithJSON(const char* json)
             }
         }
     }
-    
+
     return ret;
 }
